feat(products): Add product ratings functionality
- Introduced ProductRatingsController to handle requests for product ratings.
- Implemented GetProductRatingsUseCase to retrieve ratings and average score for a product.
- Created ProductRatingsResponseDto to structure the response data for ratings.
- Updated ProductsModule to include the new controller and use case, enhancing the product management capabilities.

These changes enable users to access and view ratings for products, improving user engagement and feedback mechanisms.
