+++ src/modules/products/products.module.ts
@@ -1,15 +1,18 @@
 import { Module } from '@nestjs/common';
 import { ProductsController } from './infrastructure/controllers/products.controller';
+import { ProductRatingsController } from './infrastructure/controllers/product-ratings.controller';
 import { ProductsRepository } from './infrastructure/repositories/products.repository';
 import { CreateProductUseCase } from './application/use-cases/create-product.use-case';
 import { UpdateProductUseCase } from './application/use-cases/update-product.use-case';
 import { DeleteProductUseCase } from './application/use-cases/delete-product.use-case';
 import { GetProductUseCase } from './application/use-cases/get-product.use-case';
 import { GetAllProductsUseCase } from './application/use-cases/get-all-products.use-case';
+import { GetProductRatingsUseCase } from './application/use-cases/get-product-ratings.use-case';
 import { PrismaModule } from '../prisma/prisma.module';
 import { UsersModule } from '../users/users.module';
 import { CloudinaryModule } from '../cloudinary/cloudinary.module';
 import { BusinessModule } from '../business/business.module';
+import { RatingsModule } from '../ratings/ratings.module';
 import { FoodTypesRepository } from './infrastructure/repositories/food-types.repository';
 import { GetAllFoodTypesUseCase } from './application/use-cases/get-all-food-types.use-case';
 import { CreateFoodTypeUseCase } from './application/use-cases/create-food-type.use-case';
@@ -20,8 +23,14 @@ import { ToggleProductPropertyUseCase } from './application/use-cases/toggle-pro
 import { ProductInventoryService } from './application/services/product-inventory.service';
 
 @Module({
-  imports: [PrismaModule, UsersModule, CloudinaryModule, BusinessModule],
-  controllers: [ProductsController],
+  imports: [
+    PrismaModule,
+    UsersModule,
+    CloudinaryModule,
+    BusinessModule,
+    RatingsModule,
+  ],
+  controllers: [ProductsController, ProductRatingsController],
   providers: [
     // Product repositories
     ProductsRepository,
@@ -36,6 +45,7 @@ import { ProductInventoryService } from './application/services/product-inventor
     DeleteProductUseCase,
     GetProductUseCase,
     GetAllProductsUseCase,
+    GetProductRatingsUseCase,
     UploadProductImageUseCase,
     ToggleProductPropertyUseCase,
