+++ src/modules/ratings/infrastructure/controllers/ratings.controller.ts
@@ -1,13 +1,4 @@
-import {
-  Controller,
-  Post,
-  Body,
-  Get,
-  Param,
-  UseGuards,
-  Query,
-  ParseIntPipe,
-} from '@nestjs/common';
+import { Controller, Post, Body, Get, Query, UseGuards } from '@nestjs/common';
 import { JwtAuthGuard } from '../../../auth/infrastructure/guards/jwt-auth.guard';
 import { RolesGuard } from '../../../auth/infrastructure/guards/roles.guard';
 import { Roles } from '../../../auth/infrastructure/decorators/roles.decorator';
@@ -15,7 +6,6 @@ import { GetUser } from '../../../auth/infrastructure/decorators/get-user.decora
 import { Role } from '@prisma/client';
 import { CreateRatingDto } from '../../application/dtos/create-rating.dto';
 import { SubmitRatingUseCase } from '../../application/use-cases/submit-rating.use-case';
-import { GetProductRatingsUseCase } from '../../application/use-cases/get-product-ratings.use-case';
 import { GetBusinessRatingsUseCase } from '../../application/use-cases/get-business-ratings.use-case';
 import { BusinessRatingsQueryDto } from '../../application/dtos/business-ratings-query.dto';
 
@@ -23,7 +13,6 @@ import { BusinessRatingsQueryDto } from '../../application/dtos/business-ratings
 export class RatingsController {
   constructor(
     private readonly submitRatingUseCase: SubmitRatingUseCase,
-    private readonly getProductRatingsUseCase: GetProductRatingsUseCase,
     private readonly getBusinessRatingsUseCase: GetBusinessRatingsUseCase,
   ) {}
 
@@ -46,17 +35,5 @@ export class RatingsController {
     @Query() query: BusinessRatingsQueryDto,
   ) {
     return this.getBusinessRatingsUseCase.execute(userId, businessId, query);
-  }
-}
-
-@Controller('products')
-export class ProductRatingsController {
-  constructor(
-    private readonly getProductRatingsUseCase: GetProductRatingsUseCase,
-  ) {}
-
-  @Get(':id/ratings')
-  async getProductRatings(@Param('id', ParseIntPipe) productId: number) {
-    return this.getProductRatingsUseCase.execute(productId);
   }
 }
